+++ src/timeouts.rs
@@ -108,8 +108,7 @@ mod test {
 
     #[test]
     fn timeout_multiplier_from_option() {
-        let args = Args::parse_from(["mutants", "--timeout-multiplier", "1.5"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
+        let options = Options::from_arg_strs(["mutants", "--timeout-multiplier", "1.5"]);
 
         assert_eq!(options.test_timeout_multiplier, Some(1.5));
         assert_eq!(
@@ -120,8 +119,8 @@ mod test {
 
     #[test]
     fn test_timeout_unaffected_by_in_place_build() {
-        let args = Args::parse_from(["mutants", "--timeout-multiplier", "1.5", "--in-place"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
+        let options =
+            Options::from_arg_strs(["mutants", "--timeout-multiplier", "1.5", "--in-place"]);
 
         assert_eq!(
             test_timeout(Some(Duration::from_secs(40)), &options),
@@ -131,8 +130,9 @@ mod test {
 
     #[test]
     fn build_timeout_multiplier_from_option() {
-        let args = Args::parse_from(["mutants", "--build-timeout-multiplier", "1.5"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
+        let args = Args::try_parse_from(["mutants", "--build-timeout-multiplier", "1.5"]).unwrap();
+        let config = Config::default();
+        let options = Options::new(&args, &config).unwrap();
 
         assert_eq!(options.build_timeout_multiplier, Some(1.5));
         assert_eq!(
@@ -143,8 +143,11 @@ mod test {
 
     #[test]
     fn build_timeout_is_affected_by_in_place_build() {
-        let args = Args::parse_from(["mutants", "--build-timeout-multiplier", "5", "--in-place"]);
-        let options = Options::new(&args, &Config::default()).unwrap();
+        let args =
+            Args::try_parse_from(["mutants", "--build-timeout-multiplier", "5", "--in-place"])
+                .unwrap();
+        let config = Config::default();
+        let options = Options::new(&args, &config).unwrap();
 
         assert_eq!(
             build_timeout(Some(Duration::from_secs(40)), &options),
@@ -154,7 +157,7 @@ mod test {
 
     #[test]
     fn timeout_multiplier_from_config() {
-        let args = Args::parse_from(["mutants"]);
+        let args = Args::try_parse_from(["mutants"]).unwrap();
         let config = Config::from_str(indoc! {r#"
             timeout_multiplier = 2.0
         "#})
@@ -170,7 +173,7 @@ mod test {
 
     #[test]
     fn build_timeout_multiplier_from_config() {
-        let args = Args::parse_from(["mutants"]);
+        let args = Args::try_parse_from(["mutants"]).unwrap();
         let config = Config::from_str(indoc! {r#"
             build_timeout_multiplier = 2.0
         "#})
@@ -186,7 +189,7 @@ mod test {
 
     #[test]
     fn timeout_multiplier_default() {
-        let args = Args::parse_from(["mutants"]);
+        let args = Args::try_parse_from(["mutants"]).unwrap();
         let options = Options::new(&args, &Config::default()).unwrap();
 
         assert_eq!(options.test_timeout_multiplier, None);
@@ -198,7 +201,7 @@ mod test {
 
     #[test]
     fn build_timeout_multiplier_default() {
-        let args = Args::parse_from(["mutants"]);
+        let args = Args::try_parse_from(["mutants"]).unwrap();
         let options = Options::new(&args, &Config::default()).unwrap();
 
         assert_eq!(options.build_timeout_multiplier, None);
@@ -207,7 +210,7 @@ mod test {
 
     #[test]
     fn timeout_from_option() {
-        let args = Args::parse_from(["mutants", "--timeout=8"]);
+        let args = Args::try_parse_from(["mutants", "--timeout=8"]).unwrap();
         let options = Options::new(&args, &Config::default()).unwrap();
 
         assert_eq!(options.test_timeout, Some(Duration::from_secs(8)));
@@ -215,7 +218,7 @@ mod test {
 
     #[test]
     fn build_timeout_from_option() {
-        let args = Args::parse_from(["mutants", "--build-timeout=4"]);
+        let args = Args::try_parse_from(["mutants", "--build-timeout=4"]).unwrap();
         let options = Options::new(&args, &Config::default()).unwrap();
 
         assert_eq!(options.build_timeout, Some(Duration::from_secs(4)));
@@ -223,7 +226,7 @@ mod test {
 
     #[test]
     fn no_default_build_timeout() {
-        let args = Args::parse_from(["mutants"]);
+        let args = Args::try_parse_from(["mutants"]).unwrap();
         let options = Options::new(&args, &Config::default()).unwrap();
 
         assert_eq!(options.build_timeout, None);
@@ -233,7 +236,7 @@ mod test {
     fn timeout_multiplier_default_with_baseline_skip() {
         // The --baseline option is not used to set the timeout but it's
         // indicative of the realistic situation.
-        let args = Args::parse_from(["mutants", "--baseline", "skip"]);
+        let args = Args::try_parse_from(["mutants", "--baseline", "skip"]).unwrap();
         let options = Options::new(&args, &Config::default()).unwrap();
 
         assert_eq!(options.test_timeout_multiplier, None);